requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Django==4.2
djangorestframework==3.15.1
pillow==10.3.0
pycodestyle==2.11.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
